+++ tests/syft/grid/duet/duet_test.py
@@ -1,19 +1,16 @@
 # stdlib
 from multiprocessing import Barrier
-from multiprocessing import Pipe
 from multiprocessing import Process
 import socket
 from time import sleep
-import traceback
-from typing import Any
 from typing import Callable
 from typing import List
-from typing import Optional
 from typing import Tuple
 
 # syft relative
 from .duet_scenarios_tests import register_duet_scenarios
 from .signaling_server_test import run
+from .syft_process_test import SyftTestProcess
 
 port = 21000
 grid_proc = Process(target=run, args=(port,))
@@ -21,27 +18,6 @@ grid_proc.start()
 
 registered_tests: List[Tuple[Callable, Callable]] = []
 register_duet_scenarios(registered_tests)
-
-
-class SyftTestProcess(Process):
-    def __init__(self, *args: Any, **kwargs: Any):
-        Process.__init__(self, *args, **kwargs)
-        self._pconn, self._cconn = Pipe()
-        self._exception = None
-
-    def run(self) -> None:
-        try:
-            Process.run(self)
-            self._cconn.send(None)
-        except Exception as e:
-            tb = traceback.format_exc()
-            self._cconn.send((e, tb))
-
-    @property
-    def exception(self) -> Optional[tuple]:
-        if self._pconn.poll():
-            self._exception = self._pconn.recv()
-        return self._exception
 
 
 def test_duet() -> None:
